fixed build without db
